# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project API",
        "version": "1.0"
    },
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "translations"
                                ],
                                "properties": {
                                    "code_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "demo_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "translations": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "language_code",
                                                "name",
                                                "description"
                                            ],
                                            "properties": {
                                                "language_code": {
                                                    "type": "string",
                                                    "maxLength": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "maxLength": 255
                                                },
                                                "description": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a project by ID",
                "operationId": "af32aed47c19d517816e0b970d60cd5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "operationId": "491d11cc6efd20335421a859159e2487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "demo_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "operationId": "187865e30ff6b72f1cea3037d9dd84fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        },
        "/api/projects/{projectId}/technologies": {
            "post": {
                "tags": [
                    "Project Technologies"
                ],
                "summary": "Assign a technology to a project",
                "operationId": "179760ccdf278230ee627dce2564dde9",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "technologie_id"
                                ],
                                "properties": {
                                    "technologie_id": {
                                        "description": "ID of the technology to assign",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Technology assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "project_id": {
                                            "type": "integer"
                                        },
                                        "technologie_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/projects/{projectId}/technologies/{technologieId}": {
            "delete": {
                "tags": [
                    "Project Technologies"
                ],
                "summary": "Remove a technology from a project",
                "operationId": "7fcb8c3b2ac98ec81effc992e6a5aed8",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "technologieId",
                        "in": "path",
                        "description": "ID of the technology to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Technology removed successfully"
                    },
                    "404": {
                        "description": "Project or Technology not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Project": {
                "required": [
                    "id",
                    "code_url",
                    "demo_url"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the project",
                        "type": "integer"
                    },
                    "code_url": {
                        "description": "The URL of the project code",
                        "type": "string"
                    },
                    "demo_url": {
                        "description": "The demo URL of the project",
                        "type": "string"
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "language_code": {
                                    "description": "Language code of the translation",
                                    "type": "string"
                                },
                                "name": {
                                    "description": "Project name in the specific language",
                                    "type": "string"
                                },
                                "description": {
                                    "description": "Project description in the specific language",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}